requirements.txt
django<1.10.0
python-social-auth<0.3.0
django-environ<0.5.0
twitter<1.18.0
django-webpack-loader<0.4.0

django-extensions<1.8.0
Werkzeug<0.12.0
